+++ src/main/java/duke/ui/Launcher.java
@@ -1,5 +1,6 @@
-package duke;
+package duke.ui;
 
+import duke.Duke;
 import javafx.application.Application;
 
 /**
